requirement.txt
numpy 
pandas 
yfinance
keras.models import load_model
streamlit
tensorflow 
matplotlib.pyplot 
sklearn.preprocessing import MinMaxScaler
sklearn.metrics import r2_score 
shapely.geometry import LineString 